refactor(file-utils): promote pathExists; replace 8 inline copies (#160 step 1)
Step 1 of #160 — consolidate the "does this path exist?" predicate.

- Move `pathExists` from `src/faiss-store-layout.ts` to
  `src/file-utils.ts` (the home that #160 nominated). The body is
  unchanged: `fsp.stat`-based, returns false for ENOENT/ENOTDIR,
  RETHROWS every other filesystem error.
- Update consumers — `faiss-store-layout.ts`, `layout-bootstrap.ts`,
  `FaissIndexManager.ts` — to import from `./file-utils.js`.

Replace 8 inline existence-check copies with `pathExists`:

  - `cli-models.ts:111, 264, 283` — three inline
    `fsp.access(p).then(() => true).catch(() => false)`. The cited
    sites in the issue.
  - `active-model.ts:82, 129, 135, 198, 230` — five
    `try { fsp.access(p); ... } catch { ... }` blocks; semantically
    the same predicate written longhand. Includes the
    `activeFileExists()` body, which is now a 1-line delegate.

Behavior change (deliberate, fix-shaped): the inline forms swallowed
ALL filesystem errors (EACCES, ELOOP, EMFILE, …) and silently
reported "does not exist". `pathExists` only swallows ENOENT/ENOTDIR
and rethrows everything else. So a permission-denied or
fd-exhaustion situation that previously looked exactly like a
missing file will now surface as a real error to the caller. The
issue called these inline copies a "future bug-fix-doesn't-propagate
trap"; this PR is what propagates the canonical behavior.

535/535 tests pass; no new test file (the predicate is exercised
transitively by 8 call sites in production code, and the existing
RFC 014 atomic-save tests already cover the lstat-vs-stat semantics
on the canonical implementation).

Refs #160 — steps 2 (consolidate the three KB-path validators), 3
(extract `assertNoTraversal`), and 4 (delete dead `resolveKbPath`)
remain as separate PRs.
