+++ src/layout-bootstrap.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import * as properLockfile from 'proper-lockfile';
 import { FAISS_INDEX_PATH } from './config.js';
 import { writeActiveModelAtomic } from './active-model.js';
-import { pathExists } from './faiss-store-layout.js';
+import { pathExists } from './file-utils.js';
 import { logger } from './logger.js';
 import { deriveModelId, type EmbeddingProvider } from './model-id.js';
